refactor: split android snapshot helper modules
